Update to the reaction logic
If a message matches the configuration filters, the handler will
increment a semaphore value and sleep. If the semaphore value is the
same when it wakes up, the handler will roll for a reaction. Otherwise,
it will exit, and allow the handler that grabbed the higher semaphore
value to resolve.